moved the component
