+++ frontend/src/components/CarDetails.tsx
@@ -5,17 +5,16 @@ import Image from "next/image";
 import { Button } from "@/components/ui/button";
 import { ArrowLeft, Verified, Loader2 } from "lucide-react";
 import { useRouter } from "next/navigation";
-import { motion } from "framer-motion";
+import { motion } from "framer-motion"; // Import motion for animations
 
-const UserCarDetail = () => {
+const CarDetails = () => {
   const [loading, setLoading] = useState(true);
-  const [isListed] = useState(false); // Listing state
   const router = useRouter();
 
   useEffect(() => {
-    // Simulate data fetching
+    // Simulate data fetching or loading delay
     const timer = setTimeout(() => setLoading(false), 1500);
-    return () => clearTimeout(timer); // Cleanup
+    return () => clearTimeout(timer); // Cleanup timer
   }, []);
 
   if (loading) {
@@ -28,14 +27,10 @@ const UserCarDetail = () => {
 
   return (
     <div className="min-h-screen bg-gray-100">
+      {/* Main Content */}
       <main className="max-w-7xl mx-auto px-4 py-8">
-        {/* Back Button with Motion Animation */}
-        <motion.div
-          initial={{ opacity: 0, y: -10 }}
-          animate={{ opacity: 1, y: 0 }}
-          transition={{ duration: 0.3 }}
-          className="mb-4"
-        >
+        {/* Back to Marketplace */}
+        <div className="mb-4">
           <Button
             variant="ghost"
             className="text-primary-action hover:underline flex items-center gap-2"
@@ -44,7 +39,7 @@ const UserCarDetail = () => {
             <ArrowLeft className="w-5 h-5" />
             Back to Dashboard
           </Button>
-        </motion.div>
+        </div>
 
         {/* Car Details Section */}
         <motion.div
@@ -54,12 +49,12 @@ const UserCarDetail = () => {
           className="bg-white shadow-lg rounded-xl overflow-hidden"
         >
           <div className="grid grid-cols-1 md:grid-cols-2 gap-8 p-8">
-            {/* Car Image with Hover Effect */}
+            {/* Car Image */}
             <motion.div
               initial={{ opacity: 0, x: -50 }}
               animate={{ opacity: 1, x: 0 }}
               transition={{ delay: 0.3, duration: 0.3 }}
-              whileHover={{ scale: 1.05 }}
+              whileHover={{ scale: 1.05, rotate: 2 }} // Hover animation (slightly enlarges and rotates)
               className="flex justify-center items-center"
             >
               <Image
@@ -67,20 +62,18 @@ const UserCarDetail = () => {
                 height={300}
                 alt="Vintage Car"
                 src="/car.jpeg"
-                className="rounded-lg object-cover transition-transform duration-300 ease-in-out"
+                className="rounded-lg object-cover transition-transform duration-300 ease-in-out" // Add transition styles
               />
             </motion.div>
 
-            {/* Car Information with Motion */}
+            {/* Car Information */}
             <motion.div
               initial={{ opacity: 0, x: 50 }}
               animate={{ opacity: 1, x: 0 }}
               transition={{ delay: 0.6, duration: 0.5 }}
               className="space-y-6"
             >
-              <h2 className="text-4xl font-bold text-gray-800">
-              Chevrolet Bel Air
-              </h2>
+              <h2 className="text-4xl font-bold text-gray-800">Chevrolet Bel Air</h2>
               <p className="text-2xl text-primary-action font-bold">$30,000</p>
               <div className="text-sm text-gray-600">
                 <p>
@@ -91,39 +84,27 @@ const UserCarDetail = () => {
                 </p>
                 <p className="flex items-center gap-1">
                   <strong>Condition:</strong> Excellent
-                  <Verified className="text-green-500 w-4 h-4" />
+                  <Verified className="text-green-500 w-4 h-4" /> {/* Updated to Verified */}
                 </p>
                 <p>
                   <strong>Service History:</strong> Up-to-date
                 </p>
-                <p
-                  className={`${
-                    isListed ? "text-green-600 bg-green-100" : "text-amber-600 bg-amber-100"
-                  } py-1 mt-2 px-4 w-[max-content]`}
-                >
-                  <strong>{isListed ? "Listed" : "Not Listed"}</strong>
-                </p>
               </div>
 
-              {/* Auction or Listing Buttons */}
+              {/* Auction or Buyout Options */}
               <div className="space-y-2">
-                {!isListed ? (
-                  <Button className="w-full bg-amber-500 text-white">
-                    List Car for Sale
-                  </Button>
-                ) : (
-                  <div className="border-t border-gray-200 pt-2">
-                    <Button className="w-full bg-amber-500 text-white">
-                      Unlist Car
-                    </Button>
-                  </div>
-                )}
+                <p className="text-lg font-semibold text-gray-700">Buyout Price: $30,000</p>
+                <Button className="w-full bg-primary-action text-white">Buy Now</Button>
+                <div className="border-t border-gray-200 pt-2">
+                  <p className="text-lg font-semibold text-gray-700">Auction Ends In: 2 days 14 hrs</p>
+                  <Button className="w-full bg-amber-500 text-white">Place a Bid</Button>
+                </div>
               </div>
             </motion.div>
           </div>
         </motion.div>
 
-        {/* Service History and Mechanic Services */}
+        {/* Car Service History and Mechanic Services */}
         <section className="mt-12">
           <motion.h3
             initial={{ opacity: 0, y: 20 }}
@@ -151,12 +132,9 @@ const UserCarDetail = () => {
           </h3>
           <div className="bg-white shadow-md rounded-lg p-4">
             <p className="text-gray-700">
-              Book a certified mechanic for repair or maintenance services
-              directly through our platform.
+              Book a certified mechanic for repair or maintenance services directly through our platform.
             </p>
-            <Button className="mt-4 bg-primary-action text-white">
-              Book a Mechanic
-            </Button>
+            <Button className="mt-4 bg-primary-action text-white">Book a Mechanic</Button>
           </div>
         </section>
       </main>
@@ -164,4 +142,4 @@ const UserCarDetail = () => {
   );
 };
 
-export default UserCarDetail;
+export default CarDetails;
